refactor(habitat): classify check diagnostics by kind
