+++ tools/habitat-harness/src/providers/grit/diagnostics.ts
@@ -6,7 +6,7 @@ import {
   observedGritDiagnosticIdentity,
   observedGritIdentityMatches,
   renderUnexpectedObservedGritIdentity,
-} from "@internal/habitat-harness/service/model/check/diagnostics/index";
+} from "@internal/habitat-harness/service/model/check/index";
 import type { RuleRunResult } from "@internal/habitat-harness/service/model/check/rule-runtime/architecture";
 import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
 import { infrastructureFailure } from "./failure.js";
