+++ tools/habitat-harness/src/providers/grit/failure.ts
@@ -1,7 +1,7 @@
 import {
   type DiagnosticProviderFailureKind,
   renderDiagnosticProviderFailure,
-} from "@internal/habitat-harness/service/model/check/diagnostics/index";
+} from "@internal/habitat-harness/service/model/check/index";
 import type { RuleRunResult } from "@internal/habitat-harness/service/model/check/rule-runtime/architecture";
 import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
