+++ tools/habitat-harness/src/providers/grit/runner.ts
@@ -6,7 +6,7 @@ import {
   type DiagnosticScanRootRefusal,
   diagnosticCatalogEntryFromRuleSourceFacts,
   renderDiagnosticScanRootRefusal,
-} from "@internal/habitat-harness/service/model/check/diagnostics/index";
+} from "@internal/habitat-harness/service/model/check/index";
 import type { RuleRunResult } from "@internal/habitat-harness/service/model/check/rule-runtime/architecture";
 import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
 import { Effect } from "effect";
