+++ tools/habitat-harness/src/providers/grit/scan-roots/index.ts
@@ -3,7 +3,7 @@ import path from "node:path";
 import {
   type DiagnosticScanRootDecision,
   renderDiagnosticScanRootRefusal,
-} from "@internal/habitat-harness/service/model/check/diagnostics/index";
+} from "@internal/habitat-harness/service/model/check/index";
 import { decideScanRootProtection } from "@internal/habitat-harness/service/model/check/protected-zones/index";
 import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
 import { repoRoot, toRepoRelative } from "@internal/habitat-harness/resources/paths";
