+++ tools/habitat-harness/src/service/model/check/dto/diagnostic-catalog.schema.ts
@@ -6,7 +6,7 @@ import {
   NativeDiagnosticIdentitySchema,
   type NativeDiagnosticIdentityValue,
   nativeDiagnosticIdentity,
-} from "./identity.js";
+} from "./diagnostic-identity.schema.js";
 
 export const GritDiagnosticScanContractSchema = Type.Object(
   {
