+++ tools/habitat-harness/src/service/model/check/dto/diagnostic-identity.schema.ts
