+++ tools/habitat-harness/src/service/model/check/dto/diagnostic-scan-root.schema.ts
