+++ tools/habitat-harness/src/service/model/check/errors/diagnostic-provider.errors.ts
