+++ tools/habitat-harness/test/lib/grit-provider.test.ts
@@ -9,7 +9,7 @@ import {
   NativeDiagnosticCatalogEntrySchema,
   observedNativeDiagnosticIdentity,
   renderDiagnosticProviderFailure,
-} from "@internal/habitat-harness/service/model/check/diagnostics/index";
+} from "@internal/habitat-harness/service/model/check/index";
 import type { RuleSourceFacts } from "@internal/habitat-harness/service/model/rules/registry/index";
 import {
   type HabitatProcessRequest,
